(REF) Extract AbstractPlusParser
